refactor(event): EventForm을 책임 단위로 나눈다
- SessionList.tsx: 세션 목록 렌더링 + 추가/수정/삭제 인라인 편집(상태 5개·mutation 3개·핸들러 5개)을 EventForm에서 분리
- rollbackEventDuplication.ts: 이벤트 복제 중 세션 생성이 실패했을 때 되돌리는 로직(deleteWithRetry, 정리 실패 항목 계산)을 분리. 처음 useEventDuplication(훅 이름)으로 만들었다가 실제로는 React Hook API를 쓰지 않아 Rules of Hooks(react-hooks/rules-of-hooks) 위반이 나서 이름·위치 변경
- EventForm.tsx: 위 두 파일로 로직이 빠지면서 594줄에서 대폭 축소. import 정리(외부/내부 구분, 알파벳 순 정렬, 중복 import 병합) 포함
- 동작 변경 없음. npm run lint·test(216개)·build 모두 통과 확인

Refs #294
